Refactor of search, better version of alpha_beta with root level pruning

+++ backend/board.c
@@ -49,7 +49,7 @@ char* board_get_fen(Board* board) {
     return board_to_fen(board);
 }
 
-float board_evaluate_current(Board* board) {
+int board_evaluate_current(Board* board) {
     return evaluate_board(board);
 }
